+++ packages/slate/test/index.js
@@ -47,12 +47,18 @@ describe('slate', () => {
   })
 })
 const withTest = editor => {
-  const { isInline, isVoid } = editor
+  const { isInline, isVoid, isElementReadOnly, isSelectable } = editor
   editor.isInline = element => {
     return element.inline === true ? true : isInline(element)
   }
   editor.isVoid = element => {
     return element.void === true ? true : isVoid(element)
+  }
+  editor.isElementReadOnly = element => {
+    return element.readOnly === true ? true : isElementReadOnly(element)
+  }
+  editor.isSelectable = element => {
+    return element.nonSelectable === true ? false : isSelectable(element)
   }
   return editor
 }